Add basic assets system imple to init assets readers (#21)
Add basic assets system imple to init assets readers.

The AssetsReaderInitSystem will async load all avaible assets readers and the caller (such as InitScene) must polling the `AssetsReaderState` to check the init finish or not.

In this submit, the InitScene doesn't check the status result but just launch the `AssetsReaderInitSystem` to load all readers.